Use same presentation in completions as hover
